# client-service/docs/openapi.yml
openapi: 3.1.0
info:
  title: Client Service API
  description: Servis za upravljanje klijentima banke. Dostupan samo zaposlenima.
  version: 1.0.0
servers:
- url: http://localhost:8083
security:
- BearerAuthentication: []
paths:
  /auth/login:
    post:
      tags:
      - auth-controller
      summary: Prijava klijenta
      description: Autentifikuje klijenta i vraca JWT token. Ne zahteva prethodni token.
      operationId: login
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequestDto"
        required: true
      responses:
        "200":
          description: Uspesna prijava
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponseDto"
        "400":
          description: Neispravni podaci u zahtevu
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Pogresan email ili lozinka
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
  /customers/{id}:
    put:
      tags:
      - client-controller
      summary: Azurira podatke klijenta
      operationId: updateClient
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdateRequestDto"
        required: true
      responses:
        "200":
          description: Klijent uspesno azuriran
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClientResponseDto"
        "400":
          description: Neispravni podaci u zahtevu
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Klijent nije pronadjen
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "409":
          description: Email adresa je vec u upotrebi
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
    delete:
      tags:
      - client-controller
      summary: Brise klijenta
      operationId: deleteClient
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Klijent uspesno obrisan
        "404":
          description: Klijent nije pronadjen
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
  /customers:
    get:
      tags:
      - client-controller
      summary: Pretrazuje klijente po filterima
      operationId: searchClients
      parameters:
      - name: ime
        in: query
        required: false
        schema:
          type: string
      - name: prezime
        in: query
        required: false
        schema:
          type: string
      - name: email
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
          maximum: 100
          minimum: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageClientResponseDto"
    post:
      tags:
      - client-controller
      summary: Kreira novog klijenta
      operationId: createClient
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientCreateRequestDto"
        required: true
      responses:
        "201":
          description: Klijent uspesno kreiran
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClientResponseDto"
        "400":
          description: Neispravni podaci u zahtevu
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "409":
          description: Email adresa ili JMBG su vec u upotrebi
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
  /customers/search:
    get:
      tags:
      - client-controller
      summary: Globalna pretraga klijenata
      operationId: globalSearchClients
      parameters:
      - name: query
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
          maximum: 100
          minimum: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageClientResponseDto"
  /customers/jmbg/{jmbg}:
    get:
      tags:
      - client-controller
      summary: Vraca ID klijenta po JMBG-u
      operationId: getIdByJmbg
      parameters:
      - name: jmbg
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClientIdResponseDto"
        "404":
          description: Klijent sa zadatim JMBG-om nije pronadjen
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
components:
  schemas:
    ClientUpdateRequestDto:
      type: object
      properties:
        ime:
          type: string
          minLength: 1
        prezime:
          type: string
          minLength: 1
        datumRodjenja:
          type: integer
          format: int64
        pol:
          type: string
          enum:
          - M
          - Z
        email:
          type: string
          format: email
        brojTelefona:
          type: string
          pattern: "^\\+?[0-9]{8,15}$"
        adresa:
          type: string
      required:
      - ime
      - prezime
    ClientResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        ime:
          type: string
        prezime:
          type: string
        datumRodjenja:
          type: integer
          format: int64
        pol:
          type: string
          enum:
          - M
          - Z
        email:
          type: string
        brojTelefona:
          type: string
        adresa:
          type: string
        role:
          type: string
          enum:
          - CLIENT_BASIC
          - CLIENT_TRADING
    ClientCreateRequestDto:
      type: object
      properties:
        ime:
          type: string
          minLength: 1
        prezime:
          type: string
          minLength: 1
        datumRodjenja:
          type: integer
          format: int64
        pol:
          type: string
          enum:
          - M
          - Z
        email:
          type: string
          format: email
          minLength: 1
        brojTelefona:
          type: string
          pattern: "^\\+?[0-9]{8,15}$"
        adresa:
          type: string
        jmbg:
          type: string
          maxLength: 13
          minLength: 13
          pattern: "^[0-9]{13}$"
      required:
      - datumRodjenja
      - email
      - ime
      - jmbg
      - pol
      - prezime
    ErrorResponseDto:
      type: object
      properties:
        errorCode:
          type: string
        errorTitle:
          type: string
        errorDesc:
          type: string
        validationErrors:
          type: object
          additionalProperties:
            type: string
    PageClientResponseDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/ClientResponseDto"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    LoginRequestDto:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
      - email
      - password
    LoginResponseDto:
      type: object
      properties:
        token:
          type: string
    ClientIdResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
  securitySchemes:
    BearerAuthentication:
      type: http
      scheme: bearer
      bearerFormat: JWT